# docs/swagger.json
{
    "swagger": "2.0",
    "info": {
        "description": "Template API for Golang project",
        "title": "Base APP Go",
        "contact": {},
        "version": "1.0"
    },
    "paths": {
...
        "/menus/{menu_id}/permissions": {
            "get": {
                "security": [
                    {
                        "BearerAuth": []
                    }
                ],
                "description": "Retrieves a paginated and searchable list of permissions (e.g., read, create) for a specific menu ID. Requires Bearer token.",
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Menu Permission"
                ],
                "summary": "Get Menu Permission List",
                "parameters": [
                    {
                        "type": "integer",
                        "description": "ID of the Menu",
                        "name": "menu_id",
                        "in": "path",
                        "required": true
                    },
                    {
                        "type": "string",
                        "description": "Search term for action_name",
                        "name": "search",
                        "in": "query"
                    },
                    {
                        "type": "integer",
                        "description": "Page number (default: 1)",
                        "name": "page",
                        "in": "query"
                    },
                    {
                        "type": "integer",
                        "description": "Items per page (default: 10)",
                        "name": "limit",
                        "in": "query"
                    },
                    {
                        "type": "string",
                        "description": "Sort order (e.g., action_name ASC)",
                        "name": "order",
                        "in": "query"
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Successfully get menu permission list",
                        "schema": {
                            "allOf": [
                                {
                                    "$ref": "#/definitions/be-dashboard-nba_internal_presentation_response.PaginatePayload"
                                },
                                {
                                    "type": "object",
                                    "properties": {
                                        "data": {
                                            "type": "array",
                                            "items": {
                                                "$ref": "#/definitions/be-dashboard-nba_internal_presentation_menu_menu-permission_presenter.ReadMenuPermissionResponse"
                                            }
                                        }
                                    }
                                }
                            ]
                        }
                    },
                    "400": {
                        "description": "Bad Request (Invalid Menu ID)",
                        "schema": {
                            "$ref": "#/definitions/be-dashboard-nba_internal_presentation_response.ErrorPayload"
                        }
                    },
                    "401": {
                        "description": "Unauthorized",
                        "schema": {
                            "$ref": "#/definitions/be-dashboard-nba_internal_presentation_response.ErrorPayload"
                        }
                    },
                    "404": {
                        "description": "Not Found (Menu ID not found)",
                        "schema": {
                            "$ref": "#/definitions/be-dashboard-nba_internal_presentation_response.ErrorPayload"
                        }
                    },
                    "500": {
                        "description": "Internal Server Error (Query parsing or usecase failure)",
                        "schema": {
                            "$ref": "#/definitions/be-dashboard-nba_internal_presentation_response.ErrorPayload"
                        }
                    }
                }
            },
            "post": {
                "security": [
                    {
                        "BearerAuth": []
                    }
                ],
                "description": "Adds a new permission (e.g., 'read', 'create') to a specific menu by its ID. Requires Bearer token.",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Menu Permission"
                ],
                "summary": "Create Menu Permission",
                "parameters": [
                    {
                        "type": "integer",
                        "description": "Menu ID",
                        "name": "menu_id",
                        "in": "path",
                        "required": true
                    },
                    {
                        "description": "Permission data to create",
                        "name": "permission",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/be-dashboard-nba_internal_presentation_menu_menu-permission_presenter.CreateMenuPermissionRequest"
                        }
                    }
                ],
                "responses": {
                    "201": {
                        "description": "Successfully create menu permission",
                        "schema": {
                            "$ref": "#/definitions/be-dashboard-nba_internal_presentation_response.DataPayload"
                        }
                    },
                    "400": {
                        "description": "Bad Request (Invalid Menu ID, Invalid JSON, or Validation Error)",
                        "schema": {
                            "$ref": "#/definitions/be-dashboard-nba_internal_presentation_response.ErrorPayload"
                        }
                    },
                    "401": {
                        "description": "Unauthorized",
                        "schema": {
                            "$ref": "#/definitions/be-dashboard-nba_internal_presentation_response.ErrorPayload"
                        }
                    },
                    "500": {
                        "description": "Internal Server Error",
                        "schema": {
                            "$ref": "#/definitions/be-dashboard-nba_internal_presentation_response.ErrorPayload"
                        }
                    }
                }
            }
        },
        "/menus/{menu_id}/permissions/{menu_permission_id}": {
            "get": {
                "security": [
                    {
                        "BearerAuth": []
                    }
                ],
                "description": "Fetches the details of a single menu permission by its unique ID. Requires Bearer token.",
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Menu Permission"
                ],
                "summary": "Get Menu Permission Detail",
                "parameters": [
                    {
                        "type": "integer",
                        "description": "Menu ID",
                        "name": "menu_id",
                        "in": "path",
                        "required": true
                    },
                    {
                        "type": "integer",
                        "description": "Menu Permission ID",
                        "name": "menu_permission_id",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Successfully get menu permission detail",
                        "schema": {
                            "allOf": [
                                {
                                    "$ref": "#/definitions/be-dashboard-nba_internal_presentation_response.DataPayload"
                                },
                                {
                                    "type": "object",
                                    "properties": {
                                        "data": {
                                            "$ref": "#/definitions/be-dashboard-nba_internal_presentation_menu_menu-permission_presenter.ReadMenuPermissionResponse"
                                        }
                                    }
                                }
                            ]
                        }
                    },
                    "400": {
                        "description": "Bad Request (Invalid ID)",
                        "schema": {
                            "$ref": "#/definitions/be-dashboard-nba_internal_presentation_response.ErrorPayload"
                        }
                    },
                    "401": {
                        "description": "Unauthorized",
                        "schema": {
                            "$ref": "#/definitions/be-dashboard-nba_internal_presentation_response.ErrorPayload"
                        }
                    },
                    "404": {
                        "description": "Not Found (Permission ID not found)",
                        "schema": {
                            "$ref": "#/definitions/be-dashboard-nba_internal_presentation_response.ErrorPayload"
                        }
                    },
                    "500": {
                        "description": "Internal Server Error",
                        "schema": {
                            "$ref": "#/definitions/be-dashboard-nba_internal_presentation_response.ErrorPayload"
                        }
                    }
                }
            },
            "put": {
                "security": [
                    {
                        "BearerAuth": []
                    }
                ],
                "description": "Updates an existing menu permission (e.g., 'code', 'action_name') by its unique ID. Requires Bearer token.",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Menu Permission"
                ],
                "summary": "Update Menu Permission",
                "parameters": [
                    {
                        "type": "integer",
                        "description": "Menu ID",
                        "name": "menu_id",
                        "in": "path",
                        "required": true
                    },
                    {
                        "type": "integer",
                        "description": "Menu Permission ID",
                        "name": "menu_permission_id",
                        "in": "path",
                        "required": true
                    },
                    {
                        "description": "Permission data to update",
                        "name": "permission",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/be-dashboard-nba_internal_presentation_menu_menu-permission_presenter.UpdateMenuPermissionRequest"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Successfully update menu permission",
                        "schema": {
                            "$ref": "#/definitions/be-dashboard-nba_internal_presentation_response.DataPayload"
                        }
                    },
                    "400": {
                        "description": "Bad Request (Invalid ID, Invalid JSON, or Validation Error)",
                        "schema": {
                            "$ref": "#/definitions/be-dashboard-nba_internal_presentation_response.ErrorPayload"
                        }
                    },
                    "401": {
                        "description": "Unauthorized",
                        "schema": {
                            "$ref": "#/definitions/be-dashboard-nba_internal_presentation_response.ErrorPayload"
                        }
                    },
                    "404": {
                        "description": "Not Found (Permission ID not found)",
                        "schema": {
                            "$ref": "#/definitions/be-dashboard-nba_internal_presentation_response.ErrorPayload"
                        }
                    },
                    "500": {
                        "description": "Internal Server Error",
                        "schema": {
                            "$ref": "#/definitions/be-dashboard-nba_internal_presentation_response.ErrorPayload"
                        }
                    }
                }
            },
            "delete": {
                "security": [
                    {
                        "BearerAuth": []
                    }
                ],
                "description": "Deletes a specific permission from a menu by its unique permission ID. Requires Bearer token.",
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Menu Permission"
                ],
                "summary": "Delete Menu Permission",
                "parameters": [
                    {
                        "type": "integer",
                        "description": "Menu ID",
                        "name": "menu_id",
                        "in": "path",
                        "required": true
                    },
                    {
                        "type": "integer",
                        "description": "Menu Permission ID",
                        "name": "menu_permission_id",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Successfully delete menu permission",
                        "schema": {
                            "$ref": "#/definitions/be-dashboard-nba_internal_presentation_response.DataPayload"
                        }
                    },
                    "400": {
                        "description": "Bad Request (Invalid ID)",
                        "schema": {
                            "$ref": "#/definitions/be-dashboard-nba_internal_presentation_response.ErrorPayload"
                        }
                    },
                    "401": {
                        "description": "Unauthorized",
                        "schema": {
                            "$ref": "#/definitions/be-dashboard-nba_internal_presentation_response.ErrorPayload"
                        }
                    },
                    "404": {
                        "description": "Not Found (Permission ID not found)",
                        "schema": {
                            "$ref": "#/definitions/be-dashboard-nba_internal_presentation_response.ErrorPayload"
                        }
                    },
                    "500": {
                        "description": "Internal Server Error",
                        "schema": {
                            "$ref": "#/definitions/be-dashboard-nba_internal_presentation_response.ErrorPayload"
                        }
                    }
                }
            }
        },
...
    },
    "definitions": {
...
        "be-dashboard-nba_internal_presentation_menu_menu-permission_presenter.CreateMenuPermissionRequest": {
            "type": "object",
            "required": [
                "action_name",
                "code"
            ],
            "properties": {
                "action_name": {
                    "type": "string",
                    "maxLength": 50,
                    "minLength": 1
                },
                "code": {
                    "type": "string",
                    "maxLength": 50,
                    "minLength": 1
                }
            }
        },
        "be-dashboard-nba_internal_presentation_menu_menu-permission_presenter.ReadMenuPermissionResponse": {
            "type": "object",
            "properties": {
                "action_name": {
                    "type": "string"
                },
                "code": {
                    "type": "string"
                },
                "id": {
                    "type": "integer"
                },
                "menu_id": {
                    "type": "integer"
                }
            }
        },
        "be-dashboard-nba_internal_presentation_menu_menu-permission_presenter.UpdateMenuPermissionRequest": {
            "type": "object",
            "required": [
                "action_name",
                "code"
            ],
            "properties": {
                "action_name": {
                    "type": "string",
                    "maxLength": 50,
                    "minLength": 1
                },
                "code": {
                    "type": "string",
                    "maxLength": 50,
                    "minLength": 1
                }
            }
        },
...
        "be-dashboard-nba_internal_presentation_response.DataPayload": {
            "type": "object",
            "properties": {
                "code": {
                    "type": "integer"
                },
                "data": {},
                "message": {
                    "type": "string"
                }
            }
        },
        "be-dashboard-nba_internal_presentation_response.ErrorPayload": {
            "type": "object",
            "properties": {
                "code": {
                    "type": "integer"
                },
                "error": {},
                "message": {
                    "type": "string"
                }
            }
        },
        "be-dashboard-nba_internal_presentation_response.PaginatePayload": {
            "type": "object",
            "properties": {
                "code": {
                    "type": "integer"
                },
                "data": {},
                "message": {
                    "type": "string"
                },
                "paginate": {
                    "$ref": "#/definitions/be-dashboard-nba_internal_presentation_response.paginatePayload"
                }
            }
        },
        "be-dashboard-nba_internal_presentation_response.paginatePayload": {
            "type": "object",
            "properties": {
                "current_page": {
                    "type": "integer"
                },
                "per_page": {
                    "type": "integer"
                },
                "total_data": {
                    "type": "integer"
                },
                "total_page": {
                    "type": "number"
                }
            }
        },
...
    },
    "securityDefinitions": {
        "BearerAuth": {
            "description": "Type \"Bearer {your token}\" into the field below.",
            "type": "apiKey",
            "name": "Authorization",
            "in": "header"
        }
    }
}